+++ src/tests/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+      "target": "ES2022",
+      "module": "ES2020",
+      "moduleResolution": "Node",
+      "removeComments": true,
+      "outDir": "../build",
+      "sourceMap": true
+    }
+  }
+
